// src/app/core/models/OriginOfGoods.model.ts
import { PaginationModel } from "./Pagination.Model";

export class OriginOfGoods extends PaginationModel {
  isSelected: boolean;
 Id: number;
 materialId: number;
 countryId: number;
 materialPercentage: number;
 locationID: number;
 comment: string;
 IsDeleted: boolean; 
 ClientId: number;
 SourceSystemId: number;
 updatedBy: string;
 updateDateTimeServer: Date;
 updateDateTimeBrowser: Date;
 updateDateTimeBrowsers: string;
  createdBy: string;
 createDateTimeBrowser: Date;
 createDateTimeBrowsers: string;
 createDateTimeServer: Date;
 id: any;
 selectRow: string;


  


}